Separate logic of 'generateErrorBasedOnCurrentEnvironment' function to separate file
